Move 'multiplayer' package to top level
