+++ src/com/reco1l/osu/Multiplayer.kt
@@ -7,7 +7,7 @@ import com.reco1l.api.ibancho.data.Room
 import com.reco1l.api.ibancho.data.RoomPlayer
 import com.reco1l.osu.conversion.jsonToScoreboardItem
 import com.reco1l.osu.conversion.jsonToStatistic
-import com.reco1l.osu.ui.multiplayer.RoomScene
+import com.reco1l.osu.multiplayer.RoomScene
 import com.reco1l.toolkt.kotlin.formatTimeMilliseconds
 import com.reco1l.toolkt.kotlin.fromDate
 import kotlinx.coroutines.CoroutineScope
