+++ src/com/reco1l/osu/multiplayer/LobbyScene.kt
@@ -1,4 +1,4 @@
-package com.reco1l.osu.ui.multiplayer
+package com.reco1l.osu.multiplayer
 
 import android.net.Uri
 import android.util.Log
