+++ src/com/reco1l/osu/multiplayer/RoomPlayerMenu.kt
@@ -1,4 +1,4 @@
-package com.reco1l.osu.ui.multiplayer
+package com.reco1l.osu.multiplayer
 
 import android.animation.Animator
 import android.app.AlertDialog
