+++ src/com/reco1l/osu/ui/MainMenu.kt
@@ -1,10 +1,10 @@
 package com.reco1l.osu.ui
 
 import com.reco1l.osu.async
-import com.reco1l.osu.ui.multiplayer.LobbyScene
+import com.reco1l.osu.multiplayer.LobbyScene
 import com.reco1l.osu.Multiplayer
 import com.reco1l.osu.beatmaplisting.BeatmapListing
-import com.reco1l.osu.ui.multiplayer.RoomScene
+import com.reco1l.osu.multiplayer.RoomScene
 import org.anddev.andengine.input.touch.TouchEvent
 import ru.nsu.ccfit.zuev.osu.LibraryManager
 import ru.nsu.ccfit.zuev.osu.MainScene
