+++ src/ru/nsu/ccfit/zuev/osu/MainActivity.java
@@ -46,8 +46,8 @@ import com.reco1l.osu.AccessibilityDetector;
 import com.reco1l.osu.Execution;
 import com.reco1l.osu.Multiplayer;
 import com.reco1l.osu.UpdateManager;
-import com.reco1l.osu.ui.multiplayer.LobbyScene;
-import com.reco1l.osu.ui.multiplayer.RoomScene;
+import com.reco1l.osu.multiplayer.LobbyScene;
+import com.reco1l.osu.multiplayer.RoomScene;
 
 import com.rian.osu.difficulty.BeatmapDifficultyCalculator;
 import net.lingala.zip4j.ZipFile;
